backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyPDF2==3.0.1
pdfplumber==0.10.3
pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.0